refactor(web-ai): split chat and settings pages
